Updated model, view, and controller for package setup
